refactor: 상태 변수명 수정
